# _tmp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SakuraFilter API",
    "description": "工业/汽车滤清器产品管理平台 API\r\n\r\n## 模块说明\r\n- **认证**: JWT 登录/刷新/登出/用户管理\r\n- **公开搜索**: 前台产品搜索/详情 (无需认证)\r\n- **后台产品管理**: CRUD (需 admin 角色)\r\n- **字典管理**: 8 类字典 CRUD (需认证)\r\n- **ETL**: 数据导入/状态/进度 (X-Admin-Token)\r\n- **运维**: 健康检查/指标/性能告警\r\n\r\n## 认证方式\r\n1. **JWT Bearer** (推荐): 登录 /api/auth/login 获取 token, 放入 Authorization: Bearer {token}\r\n2. **X-Admin-Token** (ETL/CI 备用): 放入 X-Admin-Token: {token}",
    "contact": {
      "name": "SakuraFilter Team",
      "email": "admin@sakurafilter.dev"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/storage/config": {
      "get": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "operationId": "AdminGetStorageConfig",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "operationId": "AdminSaveStorageConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageConfigDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/storage/test": {
      "post": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "operationId": "AdminTestStorageConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageConfigDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StorageConfigDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "nullable": true
          },
          "minio": {
            "$ref": "#/components/schemas/StorageEndpointConfig"
          },
          "aliyun": {
            "$ref": "#/components/schemas/StorageEndpointConfig"
          },
          "r2": {
            "$ref": "#/components/schemas/StorageEndpointConfig"
          }
        },
        "additionalProperties": false
      },
      "StorageEndpointConfig": {
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string",
            "nullable": true
          },
          "accessKey": {
            "type": "string",
            "nullable": true
          },
          "secretKey": {
            "type": "string",
            "nullable": true
          },
          "accessKeyId": {
            "type": "string",
            "nullable": true
          },
          "accessKeySecret": {
            "type": "string",
            "nullable": true
          },
          "bucketName": {
            "type": "string",
            "nullable": true
          },
          "publicEndpoint": {
            "type": "string",
            "nullable": true
          },
          "cdnEndpoint": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Admin-Token": {
        "type": "apiKey",
        "description": "Day 8.4: dev 静态 token, 从 appsettings.json 的 Auth:DevStaticToken 读",
        "name": "X-Admin-Token",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer token. 格式: Bearer {token} (登录 /api/auth/login 获取)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "X-Admin-Token": [ ]
    },
    {
      "Bearer": [ ]
    }
  ],
...
}